Implement MVVM with Android architecture libraries
The MVVM approach will make the app better maintainable and easier to
extend. The previous approach using a bound service did not scale very
well and it was basically impossible to add network synchronization of
lists.

The major problem of the bound service was to synchronize the lifecycles
of the service and the Fragments or Activities. It was not very robust
and crashed often due to nullpointers. MVVM will hopefully solve these
issues.